refactor(marker): extract the thread-safety markers into a leaf crate
Move MaybeSend/MaybeSync and their blanket impls into nectar-marker, a
new dependency-free no_std crate gated by a plain target_arch/unsync cfg,
with the unsync feature as the single knob. nectar-primitives re-exports
the traits at both crate::marker and store; nectar-file now imports them
from nectar-marker directly. The boxed aliases and the multi_thread
build.rs emitters stay in primitives and file, which keep their own cfg
sites.
