+++ .github/workflows/lint.yml
@@ -67,7 +67,8 @@ jobs:
               run: |
                   for crate in nectar-primitives nectar-postage \
                                nectar-postage-issuer nectar-postage-usage \
-                               nectar-swarms nectar-contracts nectar-file; do
+                               nectar-swarms nectar-contracts nectar-file \
+                               nectar-marker; do
                       cargo rustc --locked -p "$crate" --lib -- -D unused_crate_dependencies
                   done
                   # nectar-mantaray by path: the editor's differential gate pins a
