+++ Cargo.toml
@@ -63,6 +63,7 @@ nectar-contracts = { version = "0.4.0", path = "crates/contracts" }
 nectar-file = { version = "0.4.0", path = "crates/file" }
 nectar-manifest = { version = "0.4.0", path = "crates/manifest" }
 nectar-mantaray = { version = "0.4.0", path = "crates/mantaray" }
+nectar-marker = { version = "0.4.0", path = "crates/marker" }
 nectar-primitives = { version = "0.4.0", path = "crates/primitives" }
 nectar-swarms = { version = "0.4.0", path = "crates/swarms" }
 nectar-postage = { version = "0.4.0", path = "crates/postage" }
@@ -108,8 +109,6 @@ web-time = "1.1"
 
 # concurrency and parallelism
 rayon = "1.10"
-
-# build-script cfg aliases for the shared target/feature gates
 
 # Core dependencies
 bytes = { version = "1.11", default-features = false }
